requirements.txt
alembic==1.7.7
Flask==2.1.2
Flask_Admin==1.6.0
Flask_Bcrypt==1.0.1
flask_htmx==0.2.0
Flask_Login==0.6.1
Flask_Mail==0.9.1
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
pycoingecko==2.2.0
PyJWT==2.3.0
python_kucoin==2.1.3
requests==2.27.1
SQLAlchemy==1.4.36
Werkzeug==2.1.2
WTForms==3.0.1
